Types.ts file with common types
