+++ src/hooks/useWindowDimensions/index.ts
@@ -1,15 +1,7 @@
 // eslint-disable-next-line no-restricted-imports
 import {Dimensions, useWindowDimensions} from 'react-native';
 import variables from '../../styles/variables';
-
-type WindowDimensions = {
-    windowWidth: number;
-    windowHeight: number;
-    isExtraSmallScreenHeight: boolean;
-    isSmallScreenWidth: boolean;
-    isMediumScreenWidth: boolean;
-    isLargeScreenWidth: boolean;
-};
+import WindowDimensions from './types';
 
 /**
  * A convenience wrapper around React Native's useWindowDimensions hook that also provides booleans for our breakpoints.
