+++ src/hooks/useWindowDimensions/types.ts
@@ -1,0 +1,10 @@
+type WindowDimensions = {
+    windowWidth: number;
+    windowHeight: number;
+    isExtraSmallScreenHeight: boolean;
+    isSmallScreenWidth: boolean;
+    isMediumScreenWidth: boolean;
+    isLargeScreenWidth: boolean;
+};
+
+export default WindowDimensions;
